FEAT: Incluindo implementação do novo padrão de casos de uso para o fluxo de usuário
Obs.: Utilização do Claude para correções e implementações, os comentários serão removidos em breve.
